push new changes cart controller

+++ app/Http/Controllers/Category/CategoriesController.php
@@ -3,10 +3,12 @@
 namespace App\Http\Controllers\Category;
 
 use App\Models\Product\Category;
+use App\Models\Product\Product;
 use Illuminate\Http\JsonResponse;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreCategoryRequest;
 use App\Http\Requests\UpdateCategoryRequest;
+use App\Repositories\Category\CategoriesRepositoryInterface;
 
 class CategoriesController extends Controller
 {
@@ -15,9 +17,9 @@ class CategoriesController extends Controller
      *
      * @return JsonResponse
      */
-    public function index(): JsonResponse
+    public function index(CategoriesRepositoryInterface $categoriesRepository): JsonResponse
     {
-        return $this->response(Category::all());
+        return $this->response($categoriesRepository->get());
     }
 
     /**
@@ -62,6 +64,10 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category): JsonResponse
     {
+        if ($category->loadCount('products') > 0){
+
+            return $this->response([], 'this object has relationship', 422);
+        }
         return $this->response($category->delete());
     }
 }
